[no-verify] Update `sappho-eval` to use new flattened `sappho-east`.
