+++ eval/src/eval/expr.rs
@@ -1,7 +1,7 @@
-use super::Eval;
+use super::{Eval, EvalV};
 use crate::scope::ScopeRef;
-use crate::{Result, ValRef};
-use sappho_east::GenExpr;
+use crate::{Result, ValRef, Value};
+use sappho_east::{GenExpr, Literal};
 
 impl<FX> Eval for GenExpr<FX>
 where
@@ -11,10 +11,22 @@ where
         use GenExpr::*;
 
         match &self {
-            Universal(x) => x.eval(scope),
-            Common(x) => x.eval(scope),
-            Recursive(x) => x.eval(scope),
+            Lit(x) => x.eval(scope),
+            Ref(x) => scope.deref(x),
+            Object(x) => x.eval(scope),
+            List(x) => x.eval(scope),
+            Let(x) => x.eval(scope),
+            Apply(x) => x.eval(scope),
+            Lookup(x) => x.eval(scope),
             Effect(x) => x.eval(scope),
         }
+    }
+}
+
+impl EvalV for Literal {
+    fn eval_val(&self, _scope: ScopeRef) -> Result<Value> {
+        Ok(match self {
+            Literal::Num(f) => Value::Num(*f),
+        })
     }
 }
