+++ eval/src/eval/recursive.rs
@@ -1,23 +1,7 @@
 use super::{Eval, EvalV};
 use crate::scope::ScopeRef;
 use crate::{List, Result, ValRef, Value};
-use sappho_east::{Application, GenExpr, LetExpr, ListForm, Lookup, RecursiveExpr};
-
-impl<FX> Eval for RecursiveExpr<FX>
-where
-    FX: Eval,
-{
-    fn eval(&self, scope: ScopeRef) -> Result<ValRef> {
-        use RecursiveExpr::*;
-
-        match self {
-            List(x) => x.eval(scope),
-            Let(x) => x.eval(scope),
-            Apply(x) => x.eval(scope),
-            Lookup(x) => x.eval(scope),
-        }
-    }
-}
+use sappho_east::{Application, GenExpr, LetExpr, ListForm, Lookup};
 
 impl<FX> EvalV for ListForm<GenExpr<FX>>
 where
